requirements.txt
numpy
scikit-learn
Pillow
streamlit
